tests/unit/test_mermaid_diagram_helpers.py
#!/usr/bin/env python3
"""
Unit tests for Mermaid diagram test helper functions.

Tests the utility functions used by test_mermaid_diagrams.py:
- IgnoreUnknownLoader (custom YAML loader)
- _flatten_nav() (navigation structure flattening)
- _to_url() (path to URL conversion)
- _pages_to_check() (de-duplication)
- _mkdocs_paths() (YAML parsing with unknown tag handling)

Run with: pytest tests/unit/test_mermaid_diagram_helpers.py -v
"""

import os
from pathlib import Path
from typing import Any, Dict, List, Union
from unittest.mock import MagicMock, patch

import yaml

# Copy helper functions from test_mermaid_diagrams.py to avoid import issues
# These are pure functions that don't depend on external state


def _flatten_nav(nav: List[Union[Dict[str, Any], str]]) -> List[str]:
    """Flatten navigation structure (copied from test_mermaid_diagrams.py)."""
    paths: List[str] = []
    for item in nav:
        if isinstance(item, str):
            paths.append(item)
        elif isinstance(item, dict):
            for _title, value in item.items():
                if isinstance(value, str):
                    paths.append(value)
                elif isinstance(value, list):
                    paths.extend(_flatten_nav(value))
                elif isinstance(value, dict):
                    paths.extend(_flatten_nav([value]))
    return paths


def _to_url(md_path: str) -> str:
    """Convert markdown path to URL (copied from test_mermaid_diagrams.py)."""
    BASE_URL = os.environ.get("DOCS_BASE_URL", "http://127.0.0.1:8001")
    if md_path.startswith("docs/"):
        rel = md_path[len("docs/") :]
    else:
        rel = md_path

    if rel == "index.md":
        return f"{BASE_URL}/"
    if rel.endswith("/index.md"):
        return f"{BASE_URL}/" + rel[: -len("index.md")]

    rel_no_ext = rel[:-3]
    return f"{BASE_URL}/" + rel_no_ext + "/"


def _pages_to_check() -> List[str]:
    """Get pages to check with de-duplication (copied from test_mermaid_diagrams.py)."""
    md_paths = _mkdocs_paths()
    seen = set()
    pages: List[str] = []
    for p in md_paths:
        if p not in seen:
            seen.add(p)
            pages.append(p)
    return pages


def _mkdocs_paths() -> List[str]:
    """Parse mkdocs.yml and extract markdown paths (copied from test_mermaid_diagrams.py)."""
    REPO_ROOT = Path(__file__).resolve().parents[2]
    MKDOCS_YML = REPO_ROOT / "mkdocs.yml"

    class IgnoreUnknownLoader(yaml.SafeLoader):
        def ignore_unknown(self, node):
            return None

    IgnoreUnknownLoader.add_constructor(None, IgnoreUnknownLoader.ignore_unknown)

    with MKDOCS_YML.open("r", encoding="utf-8") as f:
        cfg = yaml.load(f, Loader=IgnoreUnknownLoader)
    nav = cfg.get("nav", [])
    raw_paths = _flatten_nav(nav)
    md_paths = [p for p in raw_paths if isinstance(p, str) and p.endswith(".md")]
    return md_paths


class TestIgnoreUnknownLoader:
    """Test the custom YAML loader that ignores unknown tags."""

    def test_ignores_unknown_tags(self):
        """Test that unknown YAML tags are ignored."""

        class IgnoreUnknownLoader(yaml.SafeLoader):
            def ignore_unknown(self, node):
                return None

        IgnoreUnknownLoader.add_constructor(None, IgnoreUnknownLoader.ignore_unknown)

        # YAML with unknown tag
        yaml_content = """
nav:
  - Home: index.md
  - !python/name:pymdownx.superfences.fence_div_format: some_value
  - Concepts: concepts/index.md
"""
        data = yaml.load(yaml_content, Loader=IgnoreUnknownLoader)

        # Should parse without error
        assert "nav" in data
        assert len(data["nav"]) == 3
        # Unknown tag should be None or ignored
        assert data["nav"][1] is None or isinstance(data["nav"][1], dict)

    def test_preserves_standard_yaml(self):
        """Test that standard YAML parsing still works."""

        class IgnoreUnknownLoader(yaml.SafeLoader):
            def ignore_unknown(self, node):
                return None

        IgnoreUnknownLoader.add_constructor(None, IgnoreUnknownLoader.ignore_unknown)

        yaml_content = """
nav:
  - Home: index.md
  - Concepts:
      - Architecture: concepts/architecture.md
      - Pipeline: concepts/pipeline_overview.md
"""
        data = yaml.load(yaml_content, Loader=IgnoreUnknownLoader)

        assert "nav" in data
        assert len(data["nav"]) == 2
        assert data["nav"][0] == {"Home": "index.md"}
        assert "Concepts" in data["nav"][1]


class TestFlattenNav:
    """Test the _flatten_nav() function."""

    def test_simple_string_list(self):
        """Test flattening a simple list of strings."""
        nav = ["index.md", "concepts/architecture.md", "how-to/quickstart.md"]
        result = _flatten_nav(nav)
        assert result == nav

    def test_nested_dictionaries(self):
        """Test flattening nested dictionary structures."""
        nav = [
            {"Home": "index.md"},
            {
                "Concepts": [
                    "concepts/architecture.md",
                    "concepts/pipeline_overview.md",
                ]
            },
        ]
        result = _flatten_nav(nav)
        assert "index.md" in result
        assert "concepts/architecture.md" in result
        assert "concepts/pipeline_overview.md" in result

    def test_mixed_structures(self):
        """Test flattening mixed string and dictionary structures."""
        nav = [
            "index.md",
            {"Concepts": "concepts/architecture.md"},
            {
                "How-To": [
                    "how-to/quickstart.md",
                    {"Advanced": "how-to/advanced.md"},
                ]
            },
        ]
        result = _flatten_nav(nav)
        assert "index.md" in result
        assert "concepts/architecture.md" in result
        assert "how-to/quickstart.md" in result
        assert "how-to/advanced.md" in result

    def test_empty_list(self):
        """Test handling empty navigation."""
        result = _flatten_nav([])
        assert result == []

    def test_deeply_nested(self):
        """Test handling deeply nested structures."""
        nav = [
            {
                "Section1": {
                    "Subsection1": [
                        "page1.md",
                        {"Subsubsection": "page2.md"},
                    ]
                }
            }
        ]
        result = _flatten_nav(nav)
        assert "page1.md" in result
        assert "page2.md" in result


class TestToUrl:
    """Test the _to_url() function."""

    def test_standard_markdown_path(self):
        """Test converting standard markdown paths to URLs."""
        assert _to_url("concepts/architecture.md") == "http://127.0.0.1:8001/concepts/architecture/"
        assert _to_url("how-to/quickstart.md") == "http://127.0.0.1:8001/how-to/quickstart/"

    def test_paths_with_docs_prefix(self):
        """Test paths that start with docs/ prefix."""
        assert (
            _to_url("docs/concepts/architecture.md")
            == "http://127.0.0.1:8001/concepts/architecture/"
        )
        assert _to_url("docs/index.md") == "http://127.0.0.1:8001/"

    def test_index_files(self):
        """Test index.md file handling."""
        assert _to_url("index.md") == "http://127.0.0.1:8001/"
        assert _to_url("concepts/index.md") == "http://127.0.0.1:8001/concepts/"
        assert _to_url("docs/concepts/index.md") == "http://127.0.0.1:8001/concepts/"

    def test_custom_base_url(self):
        """Test with custom BASE_URL environment variable."""
        original_base = os.environ.get("DOCS_BASE_URL")
        try:
            os.environ["DOCS_BASE_URL"] = "http://example.com:9000"
            # Call _to_url which reads BASE_URL from environment
            result = _to_url("concepts/architecture.md")
            assert result == "http://example.com:9000/concepts/architecture/"
        finally:
            if original_base:
                os.environ["DOCS_BASE_URL"] = original_base
            elif "DOCS_BASE_URL" in os.environ:
                del os.environ["DOCS_BASE_URL"]


class TestPagesToCheck:
    """Test the _pages_to_check() function."""

    def test_de_duplication(self):
        """Test that duplicate paths are removed."""
        # Test the de-duplication logic directly
        md_paths = [
            "index.md",
            "concepts/architecture.md",
            "index.md",  # duplicate
            "concepts/architecture.md",  # duplicate
            "how-to/quickstart.md",
        ]
        seen = set()
        pages = []
        for p in md_paths:
            if p not in seen:
                seen.add(p)
                pages.append(p)

        assert len(pages) == 3
        assert pages == ["index.md", "concepts/architecture.md", "how-to/quickstart.md"]

    def test_order_preservation(self):
        """Test that order is preserved after de-duplication."""
        md_paths = [
            "page1.md",
            "page2.md",
            "page1.md",  # duplicate
            "page3.md",
        ]
        seen = set()
        pages = []
        for p in md_paths:
            if p not in seen:
                seen.add(p)
                pages.append(p)

        assert pages == ["page1.md", "page2.md", "page3.md"]

    def test_empty_input(self):
        """Test handling empty input."""
        md_paths = []
        seen = set()
        pages = []
        for p in md_paths:
            if p not in seen:
                seen.add(p)
                pages.append(p)

        assert pages == []


class TestMkdocsPaths:
    """Test the _mkdocs_paths() function."""

    @patch("tests.unit.test_mermaid_diagram_helpers.Path")
    def test_parses_yaml_with_unknown_tags(self, mock_path_class):
        """Test that YAML with unknown tags is parsed correctly."""
        yaml_content = """
nav:
  - Home: index.md
  - !python/name:pymdownx.superfences.fence_div_format: ignored_value
  - Concepts: concepts/architecture.md
"""
        mock_path = MagicMock()
        mock_path.open.return_value.__enter__.return_value.read.return_value = yaml_content
        mock_path.open.return_value.__exit__ = lambda *args: None
        mock_path_class.return_value.resolve.return_value.parents = [MagicMock()] * 3
        mock_path_class.return_value.resolve.return_value.parents[2] = MagicMock()
        mock_path_class.return_value.resolve.return_value.parents[2].__truediv__ = (
            lambda x, y: mock_path
        )

        # Use actual file for this test since it's simpler
        # This test validates the IgnoreUnknownLoader works
        class IgnoreUnknownLoader(yaml.SafeLoader):
            def ignore_unknown(self, node):
                return None

        IgnoreUnknownLoader.add_constructor(None, IgnoreUnknownLoader.ignore_unknown)

        data = yaml.load(yaml_content, Loader=IgnoreUnknownLoader)
        assert "nav" in data
        # Should parse without error even with unknown tag

    def test_filters_non_markdown_files(self):
        """Test that non-markdown files are filtered out."""
        nav = [
            "index.md",
            "config.yaml",
            "script.py",
            "concepts/architecture.md",
        ]
        result = _flatten_nav(nav)
        # Filter to only markdown files
        md_result = [p for p in result if isinstance(p, str) and p.endswith(".md")]
        assert all(p.endswith(".md") for p in md_result)
        assert "index.md" in md_result
        assert "concepts/architecture.md" in md_result
        assert "config.yaml" not in md_result
        assert "script.py" not in md_result

    def test_handles_missing_nav_section(self):
        """Test handling empty navigation."""
        result = _flatten_nav([])
        assert result == []
